pmdash_api/requirements.txt
fastapi==0.78.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
python-dotenv~=0.19.2
uvicorn==0.17.6
pymongo==4.1.1
python_dateutil==2.8.2